requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
pydantic>=2.6
ollama
